rename api to addon_service
