+++ addon_service/addon_provider/models.py
@@ -1,0 +1,15 @@
+from django.db import models
+
+from addon_service.utils.base_model import AddonsServiceBaseModel
+
+class AddonProvider(AddonsServiceBaseModel):
+
+    provider_name = models.CharField(null=False)
+
+    def __str__(self):
+        return self.user_guid
+
+    class Meta:
+        verbose_name = "Addon Provider"
+        verbose_name_plural = "Addon Providers"
+        app_label = "addon_service"
