+++ addon_service/authorized_storage_account/views.py
