+++ addon_service/configured_storage_addon/views.py
