+++ addon_service/osf_user/serializers.py
@@ -1,0 +1,8 @@
+from addon_service.models import OSFUser
+from rest_framework_json_api import serializers
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = OSFUser
+        fields = "__all__"
