+++ addon_service/storage_provider_settings/views.py
