--- api/credentials/serializers.py
@@ -1,8 +1,0 @@
-from api.models import OSFUser
-from rest_framework_json_api import serializers
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = OSFUser
-        fields = "__all__"
